[PATCH 02/57] Make all code fit into 80 characters
Nested if statements and so on have been condensed.

If/else statements changed in a few places, to reduce the amount
of nested code.

Where tabs are literally entered into printf statements, use \t
instead of actually typing a tab character. This is more correct.

In general, make the code a bit more linear in how it reads.