core/head: Fix a tiny typo (that introduced a bug) and a segmentation fault
Those bugs were basically introduced by typos...

First bug
---------
`head` was checking if argc was less than 2, which is incorrect because
after getopt() is ran, argc would be equal to 1 if ran this way:

    $ head file

As such, it would read standard input, then the file, which is a bug.

Second bug
----------
The file got closed after the first line was read. When trying to read
the (now closed) file it returns a segmentation fault.

Signed-off-by: Ferass El Hafidi <vitali64pmemail@protonmail.com>