more: Always initialize read_file
I'm a bit confused about the structure of the code and how the read_file
variable is used here, but in any case, the the first time is entered
and when "if (!read_file)" is checked, read_file is uninitialized. If
the garbage value happens to be 0, then the loop just sets it to be 1;
if it happens to be 2, then it just prints one line and waits for input
(the odds that it's 2 are low, but this is incorrect behavior
nevertheless); if it is something else, then it behaves like 1.

Now I'm setting it to 1, so it just reads a page no matter what.

(Also, shouldn't we be using an enumerated type?)